Convert keyboardShortcuts into a React component
